removed load_cvr object to be a function
