+++ src/unnamed_rcv_thing/ballot.py
@@ -13,9 +13,3 @@ class Ballot(BaseModel):
     ranking: list[set]
     weight: float
     voters: Optional[set[str]] = None
-
-
-# if __name__ == '__main__':
-#     ballots1 = Ballot(id=None, ranking=[{'c'}, {np.nan}, {np.nan}], weight=1.0, voters={'a'})
-#     ballots2 = Ballot(id=None, ranking=[{'c'}, {np.nan}, {np.nan}], weight=1.0, voters={'a'})
-#     print(ballots1 == ballots2)
